refactor grid to get ready for selections
