+++ packages/jupyterlab-wxyz/src/widgets/index.ts
@@ -5,6 +5,10 @@ export * from './fullscreen';
 export * from './json';
 export * from './jsonld';
 export * from './markdown';
+export * from './selectgrid';
+export * from './stylegrid';
 export * from './svg';
 export * from './template';
 export * from './yaml';
+
+export * from './models/cells';
